+++ instruments/mem_profiler.rb
@@ -1,4 +1,4 @@
-require_relative '../task-2'
+require_relative '../task_2'
 require 'benchmark'
 require 'memory_profiler'
