+++ task_2.rb
